Refactor game-screen classes
